requirements.txt
# Neuro-HyMamba Brain Tumor AI System
# Production dependencies for brain tumor detection pipeline

# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0

# Medical Imaging
monai>=1.3.0
nibabel>=5.2.0
SimpleITK>=2.3.0

# Vision Models
timm>=0.9.12

# State Space Models (Mamba)
# Note: mamba-ssm requires CUDA for compilation
# Install with: pip install mamba-ssm causal-conv1d
mamba-ssm>=2.0.0
causal-conv1d>=1.0.0
einops>=0.7.0

# Data Augmentation
albumentations>=1.3.1

# Experiment Tracking (optional)
wandb>=0.16.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.1.0
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0

# CLI
click>=8.1.0
